\name{ExampleData}
\alias{ExampleData}
\alias{dataset}
\alias{class.labels}
\docType{data}

\title{The variables in the environment variable ExampleData of the system}
\description{
  The variables in the environment variable ExampleData of the system.
}
\details{
The environment variable includes the variable \code{dataset}, \code{class.labels} etc.

}
\format{An environment variable}
\author{ Junwei Han <hanjunwei1981@163.com> Yanjun Xu <tonghua605@163.com> Haixiu Yang <yanghaixiu@ems.hrbmu.edu.cn> Chunquan Li <lcqbio@yahoo.com.cn> and Xia Li <lixia@hrbmu.edu.cn>}
\keyword{file}